requirements.txt
ansible-lint==5.4.0
jmespath==0.10.0
molecule==3.6.1
molecule-docker==1.1.0
molecule-ec2==0.4
boto3==1.21.4
PyHamcrest==2.0.3
pylint==2.12.2
pytest-testinfra==6.6.0
requests==2.27.1
yamllint==1.26.3
pyyaml==6.0
